Separate screen and widget
